Implemented separate views for playlist and album songs

+++ core/signals.py
@@ -1,4 +1,5 @@
 import os
+from django.conf import settings
 from django.db.models.signals import pre_delete
 from django.dispatch import receiver
 from .models import Song, PlayList
@@ -15,10 +16,12 @@ def song_delete(sender, instance, **kwargs):
 
 @receiver(pre_delete, sender=PlayList)
 def playlist_delete(sender, instance, **kwargs):
-    if instance.cover:
+    if instance.cover and instance.cover.path != os.path.join(settings.MEDIA_ROOT, 'Default.png'):
         print("Playlist Cover Path:", instance.cover.path)
         if os.path.isfile(instance.cover.path):
             print("File Exists. Removing...")
             os.remove(instance.cover.path)
         else:
             print("File does not exist")
+    else:
+        print("Default image. Skipping deletion...")
